added repositories and refactored entities - 2

+++ src/main/java/com/example/logisticstrackingservice/repository/AuditLogRepository.java
@@ -3,5 +3,5 @@ package com.example.logisticstrackingservice.repository;
 import com.example.logisticstrackingservice.entity.AuditLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuditLogRepository extends JpaRepository<AuditLog, Integer> {
+public interface AuditLogRepository extends JpaRepository<AuditLog, Long> {
 }
