+++ src/main/java/com/example/logisticstrackingservice/repository/VehicleDriverAssignmentRepository.java
@@ -3,5 +3,5 @@ package com.example.logisticstrackingservice.repository;
 import com.example.logisticstrackingservice.entity.VehicleDriverAssignment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehicleDriverAssignmentRepository extends JpaRepository<VehicleDriverAssignment, Integer> {
+public interface VehicleDriverAssignmentRepository extends JpaRepository<VehicleDriverAssignment, Long> {
 }
